Suppress a compiler warning (int =/= uint)

+++ src/Utils/SignalDef.h
@@ -70,7 +70,7 @@ bool isCCWithFS(FitEvent *event, int nuPDG, int const (&pdgs)[N],
     for(size_t p_it2 = 0; p_it2 < N; ++p_it2) {
       nEntries += (pdgs[p_it] == pdgs[p_it2]);
     }
-    if (event->NumFSParticle(pdgs[p_it]) != nEntries) return false;
+    if ((size_t)event->NumFSParticle(pdgs[p_it]) != nEntries) return false;
   }
   return true;
 }
